Ajuste para visualizar Ranking en el historial de servicios y socios a solicitar. Calificación de servicio al finalizar. Ciclo de atención de un servicio. Se agregaron iconos necesarios para ajustes visuales. Ranking del socio.

+++ Core/MPS.Core.SharedAPIModel/Clientes/ClienteSolicitud.cs
@@ -8,7 +8,7 @@ namespace MPS.SharedAPIModel.Clientes
     {
         public Guid GUID_SOLICITUD { get; set; }
         public string FOLIO_SOLICITUD { get; set; }
-        public decimal? VALORACION_CLIENTE { get; set; }
+        public double? VALORACION_CLIENTE { get; set; }
         public string TIPO_SERVICIO { get; set; }
         public string NO_SOCIO { get; set; }
         public string SOCIO { get; set; }
